fix(review): resolve full-codebase review findings + follow-up cleanup
Max-effort full-codebase review → 14 confirmed bugs fixed (5 candidates
verified false and dropped), plus a behavior-preserving cleanup pass.
18 new tests; build + 1114 tests green; no new clippy warnings.

Fixes:
- editor: undo/redo clear the completion cycle — Ctrl-Z then Tab no
  longer replace_range's past the shrunken buffer and panics
- batch: route --batch through the safety guards (Block refuses; Confirm
  needs the new --yes), not just read_only/statement_timeout
- grid: reset_grid_view clears stale source (no cross-DB INSERT via `I`)
  and bookmarks on every new grid
- safety: dollar-quote-aware split_statements/strip_sql_comments — a `;`
  in a CREATE FUNCTION body no longer shatters/over-guards it
- reconnect: clear stuck query_running in start_connect
- schema: re-fetch the cache after editor DDL (deferred to commit for an
  auto-tx-wrapped DDL) via a new SchemaRefreshed message
- bookmarks: now per-tab (TabSnapshot) and cleared on grid replacement
- ui: centered_pct computes in u32 — no overflow/garbage overlay on
  713+ column terminals
- panels: preserve the sessions/slow-queries cursor across auto-refresh
- explain: fix the plain-EXPLAIN error-position off-by-one (23 -> 22)
- subst: track double-quoted identifiers so `?`/`$N` in a quoted column
  name isn't treated as a placeholder
- tap/otlp: derive the absolute timestamp from start + capped duration
- conn: stop logging the raw ssh_tunnel param value
- query/from_parse: remove a dead no-op loop

Cleanup (behavior-preserving; verified by the test + snapshot suite):
- ListCursorNav applied via a CursorAt adapter + apply_list_nav_key
  across the 7 tap-monitor handlers + row-detail/result-diff/cell-detail
- parallelise the bootstrap — schema::fetch's four catalog queries, and
  the version/grid/schema fetch — with tokio::join! (pipelined)
- memoise the editor highlight spans (keyed on buffer, cleared on schema
  change) so the lex+classify no longer re-runs every frame

+++ src/ui.rs
@@ -1534,8 +1534,12 @@ fn format_duration(micros: u64) -> String {
 
 /// A centred `w`%×`h`% rectangle within `area`.
 fn centered_pct(area: Rect, w: u16, h: u16) -> Rect {
-    let width = area.width * w / 100;
-    let height = area.height * h / 100;
+    // Compute in u32: `area.width * w` overflows u16 once the terminal is
+    // ~713+ columns wide (e.g. 713 × 92% > 65535), which panics in debug and
+    // wraps to a garbage rect in release. Both operands are tiny, so u32 is
+    // always lossless and the result fits back in u16.
+    let width = (area.width as u32 * w as u32 / 100) as u16;
+    let height = (area.height as u32 * h as u32 / 100) as u16;
     Rect {
         x: area.x + (area.width - width) / 2,
         y: area.y + (area.height - height) / 2,
